Add some documentation, move SampleClass
